Allow for use of absolute paths in gostatic -i.
Currently the path given by the user is always appended to the CWD. This
change allows specifying an absolute directory, in case the CWD is not
controlled by the user, or it makes sense to specify an absolute path
anyway.

Tested manually two cases cases (-i relative, -i /tmp/absolute). The
code hints at the possibility of a third case (-i/--init without a
parameter that would just use the current working directory as the
target), but I can't tell how to use that.